rearrange highlighting accessibility tests
